requirements.txt
amqp==2.6.0
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
cached-property==1.5.1
celery==4.4.5
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-admin-view-permission==1.9
django-allauth==0.41.0
django-celery==3.3.1
django-celery-beat==2.0.0
django-cors-headers==3.2.0
django-countries==5.5
django-crispy-forms==1.9.0
django-drf-auth==0.1
django-import-export==2.0.2
django-international==0.0.6
django-money==1.1
django-phonenumber-field==4.0.0
django-postgres-dropdb==2019.8.24
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docker==4.2.0
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-writable-nested==0.5.4
e==1.4.5
et-xmlfile==1.0.1
fcm-django==0.3.4
freeze==1.0.10
future==0.18.2
googletrans==2.4.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jdcal==1.4.1
jsonschema==3.2.0
kombu==4.6.10
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupPy==1.14
mccabe==0.6.1
numpy==1.18.4
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
packaging==16.8
pandas==1.0.3
paramiko==2.7.1
phonenumbers==8.12.1
Pillow==6.2.1
pkg-resources==0.0.0
postgres==3.0.0
psycopg2==2.7.3.2
psycopg2-binary==2.8.5
psycopg2-pool==1.1
py-moneyed==0.8.0
pycparser==2.20
pyfcm==1.4.7
pyfiglet==0.8.post1
PyJWT==1.7.1
pylint==2.5.2
PyNaCl==1.3.0
pyparsing==2.2.0
pyrsistent==0.16.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytrends==4.7.2
pytz==2019.3
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.1
six==1.13.0
soupsieve==2.0.1
sqlparse==0.3.0
tablib==1.1.0
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
websocket-client==0.57.0
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
zipp==3.1.0
